Change tests and CI/CD pipeline scripts
- Make tests work with evironment variables
- Refactore test config object to test environment object
- Run only api tests with the [test] target specified in the Makefile
- Make buildspec.yml read variables from the AWS param store
- Change shared-jenkins library verion to invlude vpcConfig changes
- Use vpcConfig settings in the CI pipeline
