+++ devops/aws-codebuild/Jenkinsfile.ci
@@ -1,6 +1,6 @@
 #!groovy
 
-def evLibrary = library(identifier: 'evernym-aws-codebuild@aws-codebuild', retriever: modernSCM(
+def evLibrary = library(identifier: 'evernym-aws-codebuild@v0.8.0-aws-codebuild', retriever: modernSCM(
     github(credentialsId: 'evernym-github-machine-user', repoOwner: 'evernym', repository: 'jenkins-shared')
 )).com.evernym.pipeline
 
@@ -10,6 +10,9 @@ logger.setGlobalLevel('TRACE')
 
 pipelineWrapper({
     nodeWrapper {
+
+        // Use default AWS VPC config for CodeBuild
+        env.USE_VPC_CONFIG = true
 
         List osnames = [
             'xenial',
