+++ libmysqlstorage/tests/test_utils/api_requests.rs
@@ -4,7 +4,7 @@ extern crate serde_json;
 use mysqlstorage::api as api;
 use mysqlstorage::errors::error_code::ErrorCode;
 
-use test_utils::config::{ConfigType, Config};
+use test_utils::test_env::TestEnv;
 
 use std::ffi::{CString};
 
@@ -17,14 +17,14 @@ pub mod api_requests {
     const RECORD_TYPE: &'static str = "test-type";
 
     lazy_static! {
-        static ref TEST_CONFIG: Config = Config::new(ConfigType::QA);
+        static ref TEST_ENV: TestEnv = TestEnv::new();
     }
 
 
     pub fn open_storage(wallet_name: &String) -> i32 {
         let name = CString::new(wallet_name.clone()).unwrap();
-        let config = CString::new(TEST_CONFIG.get_config()).unwrap();
-        let credentials = CString::new(TEST_CONFIG.get_credentials()).unwrap();
+        let config = CString::new(TEST_ENV.get_config()).unwrap();
+        let credentials = CString::new(TEST_ENV.get_credentials()).unwrap();
         let mut handle: i32 = -1;
 
         let err = api::open_storage(name.as_ptr(), config.as_ptr(), credentials.as_ptr(), &mut handle);
@@ -42,8 +42,8 @@ pub mod api_requests {
     pub fn create_wallet(wallet_name: &String) {
         let name = CString::new(wallet_name.clone()).unwrap();
 
-        let config = CString::new(TEST_CONFIG.get_config()).unwrap();
-        let credentials = CString::new(TEST_CONFIG.get_credentials()).unwrap();
+        let config = CString::new(TEST_ENV.get_config()).unwrap();
+        let credentials = CString::new(TEST_ENV.get_credentials()).unwrap();
         let metadata = CString::new("metadata").unwrap();
 
         let err = api::create_storage(name.as_ptr(), config.as_ptr(), credentials.as_ptr(), metadata.as_ptr());
@@ -70,8 +70,8 @@ pub mod api_requests {
 
     pub fn delete_wallet(wallet_name: &String){
         let name = CString::new(wallet_name.clone()).unwrap();
-        let config = CString::new(TEST_CONFIG.get_config()).unwrap();
-        let credentials = CString::new(TEST_CONFIG.get_credentials()).unwrap();
+        let config = CString::new(TEST_ENV.get_config()).unwrap();
+        let credentials = CString::new(TEST_ENV.get_credentials()).unwrap();
 
         let err = api::delete_storage(name.as_ptr(), config.as_ptr(), credentials.as_ptr());
         assert_eq!(err, ErrorCode::Success);
