+++ libmysqlstorage/tests/test_utils/mod.rs
@@ -1,3 +1,3 @@
-pub mod config;
+pub mod test_env;
 pub mod helper_functions;
 pub mod api_requests;
